Casted all sizeof statements to long, which matches the data type returned by the width() and height() methods of array2d. This eliminates a compiler warning when compiler with mingw
git-svn-id: file:///home/rick/.svn/richdem@252 c4c8cafc-911d-468f-9f86-3f4f8cd09a06

+++ d8_methods.h
@@ -58,7 +58,7 @@ int d8_FlowDir(const array2d<T> &elevations, const int x, const int y){
 
 template<class T>
 void d8_flow_directions(const array2d<T> &elevations, char_2d &flowdirs){
-	diagnostic_arg("The D8 flow directions will require approximately %ldMB of RAM.\n",elevations.width()*elevations.height()*sizeof(char)/1024/1024);
+	diagnostic_arg("The D8 flow directions will require approximately %ldMB of RAM.\n",elevations.width()*elevations.height()*((long)sizeof(char))/1024/1024);
 	diagnostic("Resizing flow directions matrix...");
 	flowdirs.resize(elevations.width(),elevations.height(),false);
 	diagnostic("succeeded.\n");
